#273 refactor(fe): 주식, 차트 타입 분리
