+++ src/frontend/apps/web/src/features/stock/model/index.ts
@@ -1,10 +1,5 @@
-export {
-  type Stock,
-  type StockChartAPIResponse,
-  type CandleChart,
-  type DefaultChart,
-  ChartType,
-} from './stock.types';
+export { type Stock, type StockChartAPIResponse } from './stock.types';
+export { type CandleChart, type DefaultChart, ChartType } from './chart.types';
 export { columns } from './stocks-table.columns';
 export { dummyStockData } from './stock-chart.mock';
 export { useStockChart } from './use-stock-chart';
