+++ src/frontend/apps/web/src/features/stock/model/stock.types.ts
@@ -1,5 +1,3 @@
-import { Time } from 'lightweight-charts';
-
 export type Stock = {
   id: string;
   name: string;
@@ -18,23 +16,4 @@ export type StockChartAPIResponse = {
   lowPrice: string;
   tradingVolume: string;
   totalTradeAmount: string;
-};
-
-export enum ChartType {
-  Candlestick = 'candlestick',
-  Line = 'line',
-  Histogram = 'histogram',
-}
-
-export type CandleChart = {
-  time: Time;
-  open: number;
-  high: number;
-  low: number;
-  close: number;
-};
-
-export type DefaultChart = {
-  time: Time;
-  value: number;
 };
